Refactor into gui package and modify JavaDocs
Files in the gui package run the main application window, while files
in duke package handles user interaction (e.g. add tasks, mark tasks
as done/not done, delete tasks)

Refactoring the code improves readability and accessibility of code.

Javadocs needed to be modified as the previous commit saw the
modification of several return statements in UI and Parser.
