+++ src/main/java/duke/Duke.java
@@ -1,17 +1,19 @@
 package duke;
 
+import gui.DialogBox;
 import javafx.application.Application;
+import javafx.application.Platform;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
+
 /**
  * Encapsulates the main Duke app.
  */
@@ -33,7 +35,6 @@ public class Duke extends Application {
 
     /**
      * Creates a new Duke object.
-     *
      */
     public Duke() {
         this.ui = new Ui();
@@ -48,59 +49,43 @@ public class Duke extends Application {
 
     @Override
     public void start(Stage stage) {
-        //Step 1. Setting up required components
 
-        //The container for the content of the chat to scroll.
         scrollPane = new ScrollPane();
         dialogContainer = new VBox();
         scrollPane.setContent(dialogContainer);
-
         userInput = new TextField();
         sendButton = new Button("Send");
-
         AnchorPane mainLayout = new AnchorPane();
         mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
         scene = new Scene(mainLayout);
-
         stage.setScene(scene);
         stage.show();
 
-        //Step 2. Formatting the window to look as expected
         stage.setTitle("Duke");
         stage.setResizable(false);
         stage.setMinHeight(600.0);
         stage.setMinWidth(400.0);
-
         mainLayout.setPrefSize(400.0, 600.0);
-
         scrollPane.setPrefSize(385, 535);
         scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
         scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
         scrollPane.setVvalue(1.0);
         scrollPane.setFitToWidth(true);
 
-        // You will need to import `javafx.scene.layout.Region` for this.
         dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
         userInput.setPrefWidth(325.0);
-
         sendButton.setPrefWidth(55.0);
-
         AnchorPane.setTopAnchor(scrollPane, 1.0);
-
         AnchorPane.setBottomAnchor(sendButton, 1.0);
         AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput , 1.0);
+        AnchorPane.setLeftAnchor(userInput, 1.0);
         AnchorPane.setBottomAnchor(userInput, 1.0);
 
-        //Step 3. Add functionality to handle user input.
-        sendButton.setOnMouseClicked((event) -> {
-            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
-            userInput.clear();
-        });
+//        //Step 3. Add functionality to handle user input.
+//        sendButton.setOnMouseClicked((event) -> {
+//            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
+//            userInput.clear();
+//        });
 
         userInput.setOnAction((event) -> {
             dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
@@ -121,14 +106,13 @@ public class Duke extends Application {
     /**
      * Iteration 1:
      * Creates a label with the specified text and adds it to the dialog container.
+     *
      * @param text String containing text to add
      * @return a label with the specified text that has word wrap enabled.
      */
     private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
         Label textToAdd = new Label(text);
         textToAdd.setWrapText(true);
-
         return textToAdd;
     }
 
@@ -148,43 +132,27 @@ public class Duke extends Application {
     }
 
     /**
-     * You should have your own function to generate a response to user input.
-     * Replace this stub with your completed method.
+     * Gets the response from duke by parsing user input.
+     * @param input the user input
+     * @return the String format of the response.
      */
     public String getResponse(String input) {
         try {
-            return this.parser.parseInput(input);
+            String response =  this.parser.parseInput(input);
+            if (response.equals(ui.goodbyeMessage())) {
+                Platform.exit();
+                return response;
+            }
+            return response;
         } catch (DukeException e) {
             return ui.printMessage(e.getMessage());
         }
     }
 
-
-    /**
-     * Runs the program.
-     */
-//    public void runApp() {
-//        ui.showDuke();
-//        ui.greetingMessage();
-//        boolean enteredBye = false;
-//        while (!enteredBye) {
-//            try {
-//                String input = ui.readInput();
-//                if (input.equals("bye")) {
-//                    enteredBye = true;
-//                } else {
-//                    parser.parseInput(input);
-//                }
-//                storage.save(taskList);
-//            } catch (DukeException e) {
-//                ui.printMessage(e.getMessage());
-//            }
-//        }
-//        ui.close();
-//        ui.goodbyeMessage();
-//    }
+    public String getGreeting(){
+        return ui.greetingMessage();
+    }
 }
-
