+++ src/main/java/duke/Ui.java
@@ -24,44 +24,23 @@ public class Ui {
     }
 
     /**
-     * Prints the Duke logo.
-     */
-
-    String showDuke() {
-        String logo = " ____        _\n"
-                + "|  _ \\ _   _| | _____\n"
-                + "| | | | | | | |/ / _ \\\n"
-                + "| |_| | |_| |   <  __/\n"
-                + "|____/ \\__,_|_|\\_\\___|\n";
-        return ("Hello from\n" + logo);
-    }
-
-    /**
      * Closes the scanner.
      */
     void close() {
         scanner.close();
     }
 
-//    /**
-//     * Prints a long line
-//     */
-//    public String printLongLine() {
-//        return ("\t____________________________________________________________");
-//    }
-
     /**
      * Prints a formatted message.
      * @param s The string input to be printed.
      */
     public String printMessage(String s) {
-//        printLongLine();
         return ("\t" + s);
-//        printLongLine();
     }
 
     /**
      * Prints a greeting message to the user.
+     * @return the greeting message in String format.
      */
     public String greetingMessage() {
         return printMessage("Hello! I'm Duke\n\tWhat can I do for you?");
@@ -82,11 +61,9 @@ public class Ui {
      * @return The message that tells user that a task has been added.
      */
     public String addedTaskMessage(Task t, int size) {
-//        printLongLine();
         String s1 = ("\tGot it. I've added this task:");
         String s2 = ("\n\t" + t);
         String s3 = ("\n\tNow you have " + size + " tasks in the list.");
-//        printLongLine();
         return s1 + s2 + s3;
     }
 
@@ -94,18 +71,18 @@ public class Ui {
      * Prints a formatted message showing that a task has been deleted from the list.
      * @param t The Task object to be deleted from the list.
      * @param size The size of the list after the task has been deleted.
+     * @return the message in string format that shows the deleted task and the number of tasks left.
      */
     public String deletedTaskMessage(Task t, int size) {
-//        printLongLine();
         String s1 = ("\tNoted. I've removed this task:");
         String s2 = ("\n\t" + t);
         String s3 = ("\n\tNow you have " + size + " tasks in the list");
-//        printLongLine();
         return s1 + s2 + s3;
     }
 
     /**
      * Prints a message that is displayed if there are any matching tasks to the query.
+     * @return the message that is seen before printing found tasks.
      */
     public String findTasksMessage() {
        return ("Here are the matching tasks in your list:");
@@ -115,6 +92,7 @@ public class Ui {
      * Prints the tasks found using the query function.
      * @param taskList The TaskList object containing the ArrayList of tasks
      * @param taskNumbers The Integer ArrayList containing the task numbers to retrieve from the TaskList.
+     * @return the String notation of the tasks in the StringBuilder
      */
     public String printFoundTasks(TaskList taskList, ArrayList<Integer> taskNumbers) {
         StringBuilder sb = new StringBuilder();
@@ -129,6 +107,7 @@ public class Ui {
     /**
      * Prints a message to confirm that the task has been marked as done.
      * @param t The Task object to be marked as done.
+     * @return the message in String format that shows the task modified
      */
     public String markTaskAsDoneMessage(Task t) {
         return printMessage("Nice! I've marked this task as done:\n\t" + t);
@@ -137,6 +116,7 @@ public class Ui {
     /**
      * Prints a message to confirm that the task has been marked as incomplete.
      * @param t The Task object to be marked as incomplete.
+     * @return the message in String format that shows the task modified
      */
     public String markTaskAsIncompleteMessage(Task t) {
         return printMessage("OK, I've marked this task as not done yet:\n\t" + t);
@@ -146,15 +126,14 @@ public class Ui {
     /**
      * Prints the content of ArrayList of Task objects.
      * @param tasks The ArrayList of Tasks to be printed.
+     * @return the String format of the StringBuilder containing the tasks in the list.
      */
     public String printList(ArrayList<Task> tasks) {
-//        printLongLine();
         StringBuilder sb = new StringBuilder();
         for (int i = 0; i < tasks.size(); i++) {
             int taskNumber = i + 1;
            sb.append(String.format("\t%d. %s\n", taskNumber, tasks.get(i)));
         }
-//        printLongLine();
         return sb.toString();
     }
 }
