+++ src/main/java/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package gui;
 
 import java.io.IOException;
 import java.util.Collections;
