+++ src/main/java/gui/Main.java
@@ -1,7 +1,8 @@
-package duke;
+package gui;
 
 import java.io.IOException;
 
+import duke.Duke;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -23,6 +24,7 @@ public class Main extends Application {
             Scene scene = new Scene(ap);
             stage.setScene(scene);
             fxmlLoader.<MainWindow>getController().setDuke(duke);
+            fxmlLoader.<MainWindow>getController().sendGreeting();
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
